requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
click==8.1.7
fastapi==0.108.0
flake8==7.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
python-dotenv~=1.0.0